requirements.txt
django==2.1.3
django-tastypie==0.14.2
pillow=5.3.0
